fuse-adaptor: move implementation of FUSE hl operations to separate file
